endring av datatype på antallDagerPerUke felt fra kafka melding

+++ src/main/kotlin/no/nav/arbeidsgiver/tiltakhendelseaktivitetsplan/kafka/AvtaleHendelseMelding.kt
@@ -26,5 +26,5 @@ data class AvtaleHendelseMelding(
     val sistEndret: Instant,
     val veilederNavIdent: String?,
     val annullertGrunn: String?,
-    val antallDagerPerUke: Int?
+    val antallDagerPerUke: Double?
 )
